[BE] WeekReview SQL문 수정

See merge request s10-final/S10P31D210!80

+++ back/src/main/java/com/clover/recode/domain/statistics/service/StatisticsService.java
@@ -10,7 +10,7 @@ public interface StatisticsService {
 
     StatisticsListRes getStatisticsList(Authentication authentication);
 
-    Long getReviewCnt(Authentication authentication);
+    Integer getReviewCnt(Authentication authentication);
 
      List<TodayProblemDto> getReviews(Authentication authentication);
